# web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Desk API",
    "version": "1",
    "summary": "Public, read-only data on Monad traders and wallets.",
    "description": "Keyless and rate-limited. Every response is JSON; successes are wrapped in `{ data, meta }` (health excepted), errors are RFC 9457 Problem Details with a stable `code`. Only public inputs are accepted: addresses, never keys, seeds or signatures. Limits are per client IP over a sliding minute: 60/min on the default tier, 10/min on the expensive tier (history and wallets).",
    "license": { "name": "Read-only public data" }
  },
...
  "servers": [{ "url": "https://trydesk.trade" }],
...
  "paths": {
...
    "/api/v1/markets": {
      "get": {
        "operationId": "listMarkets",
        "tags": ["Markets"],
        "summary": "Every open Perpl market",
        "description": "Perpl's market list read from its public context and turned into decimal numbers: mark and previous mark, 24h change and volume, open interest, TVL, funding, the leverage cap, fees and margin. Prices are at each market's own scale (`priceDecimals`), money is in the collateral token. Closed markets are left out. Default tier; cached 10 s.",
        "responses": {
          "200": {
            "description": "The collateral token and every open market.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": {
              "application/json": {
                "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Markets" } } }] },
                "example": {
                  "data": {
                    "collateral": "AUSD",
                    "markets": [
                      { "id": 1, "name": "BTC", "mark": 81336.1, "prev": 80450.8, "change": 0.011004241101393683, "volume24h": 40231735.848404, "openInterest": 444435.90425900003, "tvl": 165893.458031, "fundingRate": 0.00004, "fundingIntervalSec": 2580, "maxLeverage": 15, "priceDecimals": 1, "sizeDecimals": 5, "makerFee": 45, "takerFee": 345, "maintenanceMargin": 2500, "isOpen": true },
                      { "id": 10, "name": "MON", "mark": 0.024766, "prev": 0.024106, "change": 0.027379075748776276, "volume24h": 455063.158205, "openInterest": 173152.47964, "tvl": 116062.974387, "fundingRate": 0, "fundingIntervalSec": 2580, "maxLeverage": 10, "priceDecimals": 6, "sizeDecimals": 0, "makerFee": 45, "takerFee": 345, "maintenanceMargin": 2000, "isOpen": true }
                    ]
                  },
                  "meta": { "asOf": "2026-09-21T04:37:15.970Z", "chain": "monad", "cached": false }
                }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": {
            "description": "Perpl's market list could not be read. Retry with backoff.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" }, "example": { "type": "https://trydesk.trade/docs#error-upstream_unavailable", "title": "Upstream unavailable", "status": 503, "detail": "Perpl's market list could not be read: Perpl HTTP 502", "instance": "/api/v1/markets", "code": "upstream_unavailable" } } }
          }
        }
      }
    },
    "/api/v1/markets/{market}/candles": {
      "get": {
        "operationId": "marketCandles",
        "tags": ["Markets"],
        "summary": "Perpl's candles for a market",
        "description": "Candles from Perpl's own market data, so every listed market has a chart from the day it lists. Prices are in the market's quote, already scaled by its `priceDecimals`; volume is in AUSD. Up to 300 candles, oldest first, at the `bar` size asked for. A name that is not an open Perpl market answers `not_found`. Default tier; cached 60 s.",
        "parameters": [
          { "$ref": "#/components/parameters/Market" },
          {
            "name": "bar", "in": "query", "required": false,
            "description": "Candle size. Case matters: minutes are lowercase, hours and days uppercase. Anything else is refused with `invalid_request`.",
            "schema": { "type": "string", "enum": ["1m", "5m", "15m", "1H", "4H", "1D"], "default": "15m" },
            "example": "1H"
          }
        ],
        "responses": {
          "200": {
            "description": "Candles for the market, oldest first.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": {
              "application/json": {
                "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/MarketCandles" } } }] },
                "example": {
                  "data": {
                    "market": "BTC", "bar": "1H",
                    "candles": [
                      { "time": 1790870400, "open": 84113.8, "high": 84582.8, "low": 84035.9, "close": 84155.1, "volume": 809881.97941 },
                      { "time": 1790874000, "open": 84155.1, "high": 85022.7, "low": 84154.5, "close": 84983.5, "volume": 400567.996151 }
                    ]
                  },
                  "meta": { "asOf": "2026-10-01T17:48:04.262Z", "chain": "monad", "cached": false }
                }
              }
            }
          },
          "400": {
            "description": "Unknown bar.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" }, "example": { "type": "https://trydesk.trade/docs#error-invalid_request", "title": "Invalid request", "status": 400, "detail": "bar must be one of 1m, 5m, 15m, 1H, 4H, 1D.", "instance": "/api/v1/markets/BTC/candles", "code": "invalid_request" } } }
          },
          "404": {
            "description": "The market name is well-formed but is not an open Perpl market.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" }, "example": { "type": "https://trydesk.trade/docs#error-not_found", "title": "Not found", "status": 404, "detail": "No candles for DOGE.", "instance": "/api/v1/markets/DOGE/candles", "code": "not_found" } } }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": {
            "description": "Perpl did not answer or refused the request. Retry with backoff.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" }, "example": { "type": "https://trydesk.trade/docs#error-upstream_unavailable", "title": "Upstream unavailable", "status": 503, "detail": "Candles could not be read: Perpl HTTP 503", "instance": "/api/v1/markets/BTC/candles", "code": "upstream_unavailable" } } }
          }
        }
      }
    },
    "/api/v1/traders/top": {
      "get": {
        "operationId": "topTraders",
        "tags": ["Traders"],
        "summary": "Top traders on Perpl by unrealised PnL",
        "description": "Every open position on every Perpl market, summed per account and ranked by unrealised PnL. Ties and empty books are ordered by position value. Default tier; cached 30 s.",
        "parameters": [
          {
            "name": "limit", "in": "query", "required": false,
            "description": "How many traders to return, 1 to 25. Out-of-range values are clamped; anything that is not a number falls back to 25.",
            "schema": { "type": "integer", "minimum": 1, "maximum": 25, "default": 25 },
            "example": 5
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked traders with their open positions.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": {
              "application/json": {
                "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/TopTraders" } } }] },
                "example": {
                  "data": {
                    "observedAt": 1789960004040,
                    "traders": [
                      {
                        "accountId": "3901", "address": "0x95D2602d30DA1179fd13274839e60345857ca648", "pnl": "9995.72", "balance": null,
                        "positions": [
                          { "market": "ETH", "marketId": 20, "side": "long", "entry": "1847.99", "mark": "2666.85", "size": "5", "collateral": "2309.98", "value": "13334.25", "pnl": "4122.97", "pnlPercent": 178.48, "leverage": 4, "entryBlock": 90179398 }
                        ]
                      }
                    ]
                  },
                  "meta": { "asOf": "2026-09-21T03:06:44.040Z", "chain": "monad", "cached": false }
                }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Upstream" }
        }
      }
    },
    "/api/v1/traders/{address}/history": {
      "get": {
        "operationId": "traderHistory",
        "tags": ["Traders"],
        "summary": "A trader's closed trades and statistics",
        "description": "Perpl's open, close, decrease and liquidation events folded into one record per account: win rate, profit factor, drawdown, hold time, leverage, a 0–100 score, style tags, a one-line summary and the last 20 closed trades. A wallet with no Perpl account answers `stats: null`. Expensive tier: 10 requests per minute; cached 60 s.",
        "parameters": [{ "$ref": "#/components/parameters/Address" }],
        "responses": {
          "200": {
            "description": "Statistics, a one-line summary and the closed trades.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": {
              "application/json": {
                "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/TraderHistory" } } }] },
                "example": {
                  "data": {
                    "address": "0xa8805F0471c128270c49aD5B8D62111c7117c9c9", "accountId": "3679",
                    "stats": {
                      "trades": 1, "wins": 1, "losses": 0, "winRate": 1, "profitFactor": null, "realised": 0.48, "grossProfit": 0.48, "grossLoss": 0, "maxDrawdown": 0,
                      "bestStreak": 1, "worstStreak": 0, "currentStreak": 1, "averageHoldSeconds": null, "averageLeverage": null, "liquidations": 0, "volume": 0, "funding": 0.23,
                      "since": 1789248594, "lastActive": 1789248594, "bestMarket": { "symbol": "SOL", "pnl": 0.483118, "count": 1 }, "worstMarket": null, "tags": [], "score": 0
                    },
                    "summary": "Trader. Wins 100% of 1 closed trades, best on SOL.", "summarySource": "figures",
                    "trades": [{ "time": 1789248594, "market": "SOL", "side": "long", "entry": null, "exit": 101.432, "pnl": 0.483118, "holdSeconds": null, "leverage": null, "liquidated": false }]
                  },
                  "meta": { "asOf": "2026-09-21T03:07:43.859Z", "chain": "monad", "cached": false }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidAddress" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Upstream" }
        }
      }
    },
    "/api/v1/identity/{address}": {
      "get": {
        "operationId": "identity",
        "tags": ["Identity"],
        "summary": "Names and avatars for an address",
        "description": ".nad names, nad.fun profiles, ENS and Farcaster, asked in that order; the first with a name wins `name` and `source`, and an avatar, bio or X handle it lacks is borrowed from the others. Only a Farcaster account that verified the address counts. A wallet with none answers nulls. Default tier; cached 60 s, resolved names remembered for a day.",
        "parameters": [{ "$ref": "#/components/parameters/Address" }],
        "responses": {
          "200": {
            "description": "The identity record.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": {
              "application/json": {
                "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Identity" } } }] },
                "example": {
                  "data": { "address": "0xd8da6bf26964af9d7eed9e03e53415d37aa96045", "name": "vitalik.eth", "source": "ens", "avatar": "https://euc.li/vitalik.eth", "bio": "hullo", "x": "VitalikButerin", "farcaster": "vitalik.eth", "perplAccount": null },
                  "meta": { "asOf": "2026-09-21T03:06:45.121Z", "chain": "monad", "cached": false }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidAddress" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Upstream" }
        }
      }
    },
    "/api/v1/wallets/{address}": {
      "get": {
        "operationId": "wallet",
        "tags": ["Wallets"],
        "summary": "What a wallet holds and did on Monad",
        "description": "Identity, holdings on every chain OKX reads, and the wallet's trade ledger on Monad with realised and unrealised PnL. An unseen wallet is indexed on the spot (up to 25 s) and continued by a worker; `ledger.status` says which. Expensive tier: 10 requests per minute; cached 60 s.",
        "parameters": [{ "$ref": "#/components/parameters/Address" }],
        "responses": {
          "200": {
            "description": "Identity, holdings and the Monad ledger.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": {
              "application/json": {
                "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Wallet" } } }] },
                "example": {
                  "data": {
                    "address": "0x95d2602d30da1179fd13274839e60345857ca648", "observedAt": 1789960007335,
                    "identity": { "address": "0x95d2602d30da1179fd13274839e60345857ca648", "name": null, "source": null, "avatar": null, "bio": null, "x": null, "farcaster": null, "perplAccount": "3901" },
                    "portfolio": null, "chains": [], "held": null, "holdings": [],
                    "ledger": {
                      "status": "ready", "indexedAt": 1789960006791, "behind": 0, "realized": -3.868253284869727, "unrealized": 0, "value": 0, "winRate": 0.6666666666666666,
                      "last7d": { "realized": 0, "trades": 2, "winRate": 1 }, "last30d": { "realized": -3.868253284869727, "trades": 6, "winRate": 0.6666666666666666 },
                      "tokens": [{ "token": "0x00000000efe302beaa2b3e6e1b18d08d69a9012a", "symbol": "AUSD", "holding": 0, "unpriced": 0, "averageCost": null, "price": 1.0022532779189006, "value": 0, "realized": -3.868253284869727, "unrealized": 0 }],
                      "trades": [{ "time": 1789797877000, "hash": "0x171fdca313496c0757fb42e093fbd2fe57fefa82367606453beb7b462adf078d", "token": "0x00000000efe302beaa2b3e6e1b18d08d69a9012a", "symbol": "AUSD", "side": "sell", "amount": 2039.77, "price": 0.9983990165892015, "value": 2036.5043620681556, "gain": 0 }]
                    },
                    "labels": [{ "code": "perpl", "text": "Trades perps here · Perpl #3901" }],
                    "logos": { "143:0x00000000efe302beaa2b3e6e1b18d08d69a9012a": "https://static.oklink.com/cdn/web3/currency/token/large/143-0x00000000efe302beaa2b3e6e1b18d08d69a9012a-105/type=default_90_0?v=1789931285453" }
                  },
                  "meta": { "asOf": "2026-09-21T03:06:47.373Z", "chain": "monad", "cached": false }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidAddress" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Upstream" }
        }
      }
    },
    "/api/v1/tokens/signals": {
      "get": {
        "operationId": "tokenSignals",
        "tags": ["Signals"],
        "summary": "Token signals from tracked wallets over a window",
        "description": "Tokens that the Perpl leaderboard, followed wallets and proven tracked wallets have been buying, scored 0–100 by distinct buyers, net inflow, first-time buys and freshness, less a penalty for concentration. At least two buyers of $100+ and a score of 50 to appear; stablecoins and MON wrappers excluded; at most 25. Default tier; cached 15 s, recomputed at most once a minute.",
        "parameters": [
          {
            "name": "window", "in": "query", "required": false,
            "description": "How far back to look. Anything else is refused with `invalid_request`.",
            "schema": { "type": "string", "enum": ["1h", "6h", "24h"], "default": "6h" },
            "example": "6h"
          }
        ],
        "responses": {
          "200": {
            "description": "Scored tokens for the window. An empty list is a real answer.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": {
              "application/json": {
                "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Signals" } } }] },
                "examples": {
                  "quiet": { "summary": "Captured live: nothing qualified in the window", "value": { "data": { "observedAt": 1789960007855, "window": "6h", "signals": [] }, "meta": { "asOf": "2026-09-21T03:06:47.885Z", "chain": "monad", "cached": false } } },
                  "populated": {
                    "summary": "One signal, in the shape the server builds",
                    "value": {
                      "data": {
                        "observedAt": 1789960007855, "window": "6h",
                        "signals": [{ "token": "0x1234567890abcdef1234567890abcdef12345678", "symbol": "CHOG", "chainIndex": "143", "score": 82, "strong": true, "summary": "3 top traders bought CHOG in the last 6 hours · +$4.2K net", "buyers": [{ "address": "0x95d2602d30da1179fd13274839e60345857ca648" }], "distinct": 3, "netUsd": 4210.5, "firstBuyShare": 0.667, "warning": null, "price": 0.0412 }]
                      },
                      "meta": { "asOf": "2026-09-21T03:06:47.885Z", "chain": "monad", "cached": false }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown window.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" }, "example": { "type": "https://trydesk.trade/docs#error-invalid_request", "title": "Invalid request", "status": 400, "detail": "window must be one of 1h, 6h, 24h.", "instance": "/api/v1/tokens/signals", "code": "invalid_request" } } }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Upstream" }
        }
      }
    },
...
    "/api/v1/stats": {
      "get": {
        "operationId": "stats",
        "tags": ["Monitoring"],
        "summary": "Usage counters",
        "description": "Counts only: requests served today, tracked wallets, alert subscriptions. No addresses. Default tier; cached 30 s.",
        "responses": {
          "200": {
            "description": "Counters.",
            "headers": { "$ref": "#/components/headers/RateLimit" },
            "content": {
              "application/json": {
                "schema": { "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Stats" } } }] },
                "example": { "data": { "day": "2026-09-21", "requestsToday": 14, "trackedWallets": 22, "alertSubscriptions": 3 }, "meta": { "asOf": "2026-09-21T03:06:49.038Z", "chain": "monad", "cached": false } }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "503": { "$ref": "#/components/responses/Upstream" }
        }
      }
    }
  },
  "components": {
...
    "headers": {
      "RateLimit": {
        "X-RateLimit-Limit": { "schema": { "type": "integer" }, "description": "Requests allowed per minute on this route's tier." },
        "X-RateLimit-Remaining": { "schema": { "type": "integer" }, "description": "Requests left in the current window; 0 when over." },
        "X-RateLimit-Reset": { "schema": { "type": "integer" }, "description": "Unix seconds when the window resets." }
      }
    },
...
    "schemas": {
...
      "Envelope": {
        "type": "object",
        "required": ["data", "meta"],
        "properties": {
          "data": { "description": "The route's payload." },
          "meta": {
            "type": "object",
            "required": ["asOf", "chain", "cached"],
            "properties": {
              "asOf": { "type": "string", "format": "date-time", "description": "When the server produced the answer." },
              "chain": { "type": "string", "const": "monad" },
              "cached": { "type": "boolean", "description": "True when served from the server's own memo rather than a fresh read." }
            }
          }
        }
      },
...
      "Markets": {
        "type": "object",
        "required": ["collateral", "markets"],
        "properties": {
          "collateral": { "type": "string", "description": "Symbol of the collateral token every money figure is in; AUSD." },
          "markets": { "type": "array", "items": { "$ref": "#/components/schemas/Market" }, "description": "Every open market, in the venue's order." }
        }
      },
...
    }
  }
}